requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dateutil>=2.8.0
tabulate>=0.9.0
fastapi>=0.119.0
uvicorn>=0.37.0
pydantic>=2.12.0
